refactor(kernel): move evidence discovery out of request_handlers
request_handlers is the dispatch table, three lane drivers, the shape
materialization and the integrate path all at once. The scanners that answer
"what did this session already produce" are none of those: which serving log
carries aiter dispatch evidence, which trace the fusion discover stage can
read, which untuned CSV belongs to this model. They are reads, and every lane
needs the same answers, so they belong in a module a lane can import rather
than in the one lane that happened to need them first.

kernel_evidence collects them. The six names a caller outside the module uses
lose their leading underscore, because a module-private spelling on a
cross-module import says the opposite of what is true. The rest stay private.

The split runs through the tests too. The server-log suite was two suites in
one file -- the scanners, and what the GEMM lane forwards once they have
answered -- so each half now sits with the module it pins. The three
server-log cases in the units file are dropped rather than moved: the
evidence suite already covers each of them and additionally requires the
dispatch evidence they predate.

No behaviour change.
